attempt at making master script that calls the rockstar script
